refactor: インフラストラクチャ層（Infrastructure）への外部API連携・DB管理の分離とファイル移動
