+++ Frontend/src/infrastructure/api/dictionaryApi.ts
